1963 Scripts for Mac checks if the environment is Mac

+++ dev/macos_finder_settings.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
 
-# スクリーンショットの撮影時に影を含めない
-defaults write com.apple.screencapture disable-shadow -boolean true
+os=$(uname)
 
-# Finder で隠しファイルを表示する
-defaults write com.apple.finder AppleShowAllFiles true
+if [ "$os" = "Darwin" ]; then
+  # スクリーンショットの撮影時に影を含めない
+  defaults write com.apple.screencapture disable-shadow -boolean true
 
-# スクリーンショットのファイル名変更
-defaults write com.apple.screencapture name "Screenshot"
+  # Finder で隠しファイルを表示する
+  defaults write com.apple.finder AppleShowAllFiles true
 
-# SystemUIServer を再起動して設定を反映させる
-killall SystemUIServer
+  # スクリーンショットのファイル名変更
+  defaults write com.apple.screencapture name "Screenshot"
 
+  # SystemUIServer を再起動して設定を反映させる
+  killall SystemUIServer
+else
+  exit 1
+fi
+
+exit 0
